requirements.txt
torch
pillow
diffusers @ git+https://github.com/huggingface/diffusers
transformers
accelerate


